Module pallet_example_kitchensink::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AlternativeSyntax

In all of these examples, we chose a syntax where the storage item is defined using the explicit generic syntax (X = Y). Alternatively:

AlternativeSyntax2

Lastly, all storage items, as you saw, had to be generic over T. If they want to use an item from Config, <T: Config> should be used.

Bar

A storage map. This creates a mapping from keys of type u32 to values of type u32.

Foo

A storage value. We mark this as unbounded, alter its prefix, and define a custom storage getter for it.

FooWithDefault

Warning: Doc-Only

Quux

Same as StorageDoubleMap, but with arbitrary number of keys.

Qux

Conceptually same as StorageMap<> where the key is a tuple of (u32, u32). On top, it provides some functions to iterate or remove items based on only the first key.